changed isolate_number's return from an int to a long

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: algautie <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/24 12:24:48 by algautie          #+#    #+#             */
-/*   Updated: 2019/06/26 10:11:53 by algautie         ###   ########.fr       */
+/*   Updated: 2019/06/26 10:22:16 by algautie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,7 +91,7 @@ char				*ft_itoa_base(int value, int base);
 void				ft_sort_integer_table(int *tab, int size);
 int					ft_strdchr(char *str, char c, int start);
 int					ft_cmpchars(char c, char *cs);
-int					ft_isolate_number(char *str, int start);
+long				ft_isolate_number(char *str, int start);
 long				ft_atolong(const char *str);
 
 #endif
